+++ Gemfile
@@ -6,10 +6,6 @@ gem 'geocoder'
 
 gem 'react_on_rails'
 
-gem 'carrierwave', '~> 1.0'
-
-
-gem "aws-sdk-s3", require: false
 
 gem 'dotenv-rails'
 # Allow Facebook signup/login
