+++ app/assets/stylesheets/about.scss
@@ -1,3 +1,3 @@
-// Place all the styles related to the Support controller here.
+// Place all the styles related to the About controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: http://sass-lang.com/
