+++ app/controllers/about_controller.rb
@@ -1,0 +1,4 @@
+class AboutController < ApplicationController
+  def index
+  end
+end
