--- app/controllers/hello_world_controller.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class HelloWorldController < ApplicationController
-  layout "hello_world"
-
-  def index
-    @hello_world_props = { name: "Stranger" }
-  end
-end
