--- app/controllers/map_controller.rb
@@ -1,8 +1,0 @@
-class MapController < ApplicationController
-  before_action :authenticate_user!
-  def googlemap
-  end
-
-  def mapbox
-  end
-end
