--- app/controllers/support_controller.rb
@@ -1,4 +1,0 @@
-class SupportController < ApplicationController
-  def index
-  end
-end
