+++ app/controllers/user_controller.rb
@@ -1,6 +1,8 @@
 class UserController < ApplicationController
+  #Require user be signed in
   before_action :authenticate_user!
   def index
+    #Renders all users then sort by  how many points they have
     respond_to do |format|
       format.html
       format.json do
