+++ app/helpers/about_helper.rb
@@ -1,0 +1,2 @@
+module AboutHelper
+end
