--- app/helpers/map_helper.rb
@@ -1,2 +1,0 @@
-module MapHelper
-end
