--- app/helpers/report_helper.rb
@@ -1,2 +1,0 @@
-module ReportHelper
-end
