--- app/helpers/support_helper.rb
@@ -1,2 +1,0 @@
-module SupportHelper
-end
