+++ app/javascript/bundles/App/App.js
@@ -1,10 +1,6 @@
 import React, { Component } from 'react';
 import Nav from './Components/NavBar.js'
 import LandingPage from './Components/Landing.js';
-import ReactDOM from 'react-dom'
-import 'font-awesome/css/font-awesome.min.css';
-import 'bootstrap/dist/css/bootstrap.min.css';
-import 'mdbreact/dist/css/mdb.css';
 import ModalPage from './Components/Modal';
 
 class App extends Component {
