+++ app/javascript/bundles/App/Components/Maps/Map.js
@@ -2,14 +2,19 @@ import React, { Component } from 'react';
 import mapboxgl from 'mapbox-gl';
 import axios from 'axios';
 
+
 export default class Map extends Component {
 
+//Inherits window.map from wherever its called
   constructor(){
     super();
     window.map = this;
   }
 
+//METHODS AND FUNCTIONS THAT WILL TAKE PLACE AFTER COMPONENT MOUNTS THE DOM
   componentDidMount() {
+
+//API KEY FOR MAPBOX
     mapboxgl.accessToken = 'pk.eyJ1IjoiYW5keXdlaXNzMTk4MiIsImEiOiJIeHpkYVBrIn0.3N03oecxx5TaQz7YLg2HqA'
 
     let map;
@@ -19,19 +24,25 @@ export default class Map extends Component {
     let geo;
     let { coordinates, geolocate } = this.props;
 
+    //OPTIONS FOR BUILT IN GEOLOCATOR BUTTON
     const geolocationOptions = {
+    //Tells Geocoder to use gps locating over ip locating
       enableHighAccuracy: true,
+    //Sets maximum wait time
       maximumAge        : 30000,
       timeout           : 27000
     };
 
+    //OPTIONS FOR MAPBOX COMPONENT
     const mapOptions = {
+      //DEFINES CONTAINER
       container: this.mapContainer,
       style: `mapbox://styles/mapbox/streets-v9`,
       zoom: 12,
       center: [-80.2044, 25.8028]
     }
 
+    //IF GEOLOCATION IS ACTIVE THEN CENTER MAP AT CURRENT LOCATION
     if ("geolocation" in navigator && geolocate) {
       navigator.geolocation.getCurrentPosition(
         // success callback
@@ -52,39 +63,44 @@ export default class Map extends Component {
     }
   }
 
-  geo = new MapboxGeocoder({
-    accessToken: mapboxgl.accessToken
-  })
-
+  //INITIALIZE MAPS
   createMap = (mapOptions, geolocationOptions) => {
     this.map = new mapboxgl.Map(mapOptions);
     const map = this.map;
+    //CENTERS MAP - REFER TO MAP-OPTIONS
     const { lat, lng } = map.getCenter();
     console.log(lat, lng);
+    //APPENDS SEARCH BAR NAVIGATOR
     map.addControl(
       new MapboxGeocoder({
         accessToken: mapboxgl.accessToken
       })
     );
+    //APPENDS GEOLOCATOR BUTTON
     map.addControl(
       new mapboxgl.GeolocateControl({
         positionOptions: geolocationOptions,
         trackUserLocation: true
       })
     );
+    //APPEND EASY ZOOM IN / ZOOM OUT CONTROLS
     map.addControl(
       new mapboxgl.NavigationControl({
         positionOptions: geolocationOptions,
         trackUserLocation: true
       })
     );
+    //ON MAP LOAD, ADD ALL PLACE MARKERS FROM .JSON DATA
     map.on('load', (event) => {
       map.addSource(
         'places',
         { type: 'geojson', data: `/places.json?lat=${lat}&lng=${lng}` }
       );
+      //ADD MARKERS TO MAP
       map.addLayer({ id: 'places', type: 'circle', source: 'places'});
+      //AFTER MAP SETTLES, FETCH NEW PLACE
       map.on('moveend', (e) => { this.fetchPlaces() });
+      //SHOW POPUP ON CLICK -- STILL NEEDS TO BE STYLED
       map.on('click', (e) => {
         const features = map.queryRenderedFeatures(e.point, { layers: ['places'] });
         if (!features.length) { return; }
@@ -94,15 +110,15 @@ export default class Map extends Component {
                         .setHTML(`<a href='/places/${feature.properties['id']}'>${feature.properties['name']}</a>`)
                         .addTo(map);
       });
+      //IF MOUSE MOVES ONTOP A POPUP, CHANGE CURSOR TYPE
       map.on('mousemove', (e) => {
         const features = map.queryRenderedFeatures(e.point, { layers: ['places'] });
         map.getCanvas().style.cursor = features.length ? 'pointer' : '';
       });
-
-
     });
   }
 
+  //METHOD THAT MAKES AXIOS REQUEST FOR PLACES.JSON
   fetchPlaces = () => {
     const map = this.map;
     const { lat, lng } = map.getCenter();
@@ -111,6 +127,7 @@ export default class Map extends Component {
       .catch((error) => { console.log(error) });
   }
 
+  //ACTION FOR WHEN COMPONENT LEAVES THE DOM -- UNSAFE?
   componentWillUnmount() {
     this.map.remove();
   }
