+++ app/javascript/bundles/App/Components/Notes/Form.js
@@ -3,7 +3,6 @@ import Paper from '@material-ui/core/Paper';
 import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
 import Grid from '@material-ui/core/Grid';
-import axios from 'axios';
 
 
 export default class Form extends Component {
