+++ app/javascript/bundles/App/Components/Notes/InitMap.js
