+++ app/javascript/bundles/App/Components/Scoreboard.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import axios from 'axios';
 import Paper from '@material-ui/core/Paper';
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
