+++ app/javascript/bundles/App/Components/SideNav.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { BrowserRouter as Router } from 'react-router-dom';
 import { Fa, SideNavItem, SideNavCat, SideNavNav, SideNav, SideNavLink } from 'mdbreact';
 
 class SideNavPage extends React.Component {
