+++ app/javascript/packs/App.js
@@ -1,7 +1,6 @@
 import ReactOnRails from 'react-on-rails';
 import ReactDOM from 'react-dom'
 import App from '../bundles/App/App.js';
-import Report from '../bundles/App/Components/Report.js'
 import Places from '../bundles/App/Components/Maps/Places.js'
 import Contact from '../bundles/App/Components/Contact.js'
 import About from '../bundles/App/Components/About.js'
@@ -10,7 +9,7 @@ import Savedspots from '../bundles/App/Components/Savedspots.js'
 
 
 ReactOnRails.register({
-
-    App, Report, Places, Contact, About, Scoreboard, Savedspots
+  //REGISTER COMPONENTS BEFORE USING!
+    App, Places, Contact, About, Scoreboard, Savedspots
 
 });
