+++ app/models/application_record.rb
@@ -1,3 +1,4 @@
 class ApplicationRecord < ActiveRecord::Base
+  #Whatever is put here will apply to all models 
   self.abstract_class = true
 end
