--- app/models/report.rb
@@ -1,4 +1,0 @@
-class Report < ApplicationRecord
-  belongs_to :user
-  has_many :votes
-end
