+++ app/models/user.rb
@@ -1,7 +1,8 @@
 class User < ApplicationRecord
+  #Relationships the user has with other modles
   has_many :votes
-  has_many :reports
   has_many :places
+  #Sorts users by points
   scope :by_points, -> { order(points: :desc)}
 
   # Include default devise modules. Others available are:
@@ -10,6 +11,7 @@ class User < ApplicationRecord
          :recoverable, :rememberable, :trackable, :validatable
   devise :omniauthable, omniauth_providers: [:facebook]
 
+  #Facebook login?
   def self.from_omniauth(auth)
     where(email: auth.info.email).first_or_create! do |user|
       user.provider = auth.provider
