+++ app/models/vote.rb
@@ -1,5 +1,4 @@
 class Vote < ApplicationRecord
   belongs_to :user
-  belongs_to :report
   belongs_to :place
 end
