--- app/views/hello_world/index.html.erb
@@ -1,2 +1,0 @@
-<h1>Hello World</h1>
-<%= react_component("HelloWorld", props: @hello_world_props, prerender: false) %>
