--- app/views/layouts/hello_world.html.erb
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>ReactOnRailsWithWebpacker</title>
-    <%= csrf_meta_tags %>
-    <%= javascript_pack_tag 'hello-world-bundle' %>
-  </head>
-
-  <body>
-    <%= yield %>
-  </body>
-</html>
