+++ app/views/places/index.html.erb
@@ -1,6 +1,3 @@
-Hello From index
-
-You are calling from
 
 <%= @coordinates.first %>, <%= @coordinates.last %>
 
@@ -32,14 +29,8 @@ You are calling from
   <div class="field">
     <%= form.label :image %>
     <%= form.file_field :image %>
-    <% if @place.image.attached? %>
-      <p>
-        <strong>Image:</strong>
-        <br>
-        <%= image_tag @place.image %>
-      </p>
-      <% end %>
   </div>
+  <div>
   <div>
     <%= form.submit %>
   </div>
