--- app/views/report/index.html.erb
@@ -1,3 +1,0 @@
-<%= react_component("Report", props: { user: current_user }, prerender: false) %>
-<%= link_to 'Home', '/' %>
-<%= link_to 'Sign Out', destroy_user_session_path, method: :delete %>
