--- db/migrate/20180719190923_add_user_to_reports.rb
@@ -1,5 +1,0 @@
-class AddUserToReports < ActiveRecord::Migration[5.2]
-  def change
-    add_reference :reports, :user, foreign_key: true
-  end
-end
