--- db/migrate/20180719191014_add_report_to_votes.rb
@@ -1,5 +1,0 @@
-class AddReportToVotes < ActiveRecord::Migration[5.2]
-  def change
-    add_reference :votes, :report, foreign_key: true
-  end
-end
