--- db/migrate/20180723055718_change_lat_and_lng_to_be_float_in_reports.rb
@@ -1,6 +1,0 @@
-class ChangeLatAndLngToBeFloatInReports < ActiveRecord::Migration[5.2]
-  def change
-    change_column :reports, :lat, :float
-    change_column :reports, :lng, :float
-  end
-end
