+++ db/schema.rb
@@ -52,18 +52,6 @@ ActiveRecord::Schema.define(version: 2018_07_25_042634) do
     t.index ["user_id"], name: "index_places_on_user_id"
   end
 
-  create_table "reports", force: :cascade do |t|
-    t.string "description"
-    t.float "lat"
-    t.float "lng"
-    t.string "photo"
-    t.string "category"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.integer "user_id"
-    t.index ["user_id"], name: "index_reports_on_user_id"
-  end
-
   create_table "users", force: :cascade do |t|
     t.string "email", default: "", null: false
     t.string "encrypted_password", default: "", null: false
@@ -90,10 +78,8 @@ ActiveRecord::Schema.define(version: 2018_07_25_042634) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.integer "user_id"
-    t.integer "report_id"
     t.integer "place_id"
     t.index ["place_id"], name: "index_votes_on_place_id"
-    t.index ["report_id"], name: "index_votes_on_report_id"
     t.index ["user_id"], name: "index_votes_on_user_id"
   end
